==> notebooks/1.0-nlg-load-grammys-to-db.ipynb <==
# Workshop #2 :
## Load - Grammys Dataset to DB

Load historical Grammy Awards data from CSV into PostgreSQL.

import sys
import os
import logging
import pandas as pd
from sqlalchemy import create_engine
from sqlalchemy.sql import text
from sqlalchemy.exc import SQLAlchemyError


sys.path.append(os.path.abspath('../'))
from src.params import Params
from src.client import DatabaseClient
from src.logging_config import setup_logging

setup_logging()

# Path to the Grammy's dataset
csv_file = '../data/external/the_grammy_awards.csv'

df = pd.read_csv(csv_file)

df.head()

A Parameters class centralizes anything that can be parametrized in the code. As we want to use the parameters for the connection to the PostgrSQL databases, params are instantiated as `params = Params()`.

# Parames instance
params = Params()

# Connection to the database
db_client = DatabaseClient(params)

A `table_name` for the grammy's dataset table is defined and then the Dataframe is writed to the PostgreSWL with SQLAlchemy's mehtod `df.to.sql()`

table_name = 'grammys_raw'

try:
    # Write the DataFrame to the database table
    df.to_sql(table_name, con=db_client.engine, if_exists='replace', index=False)
    logging.info(f"CSV data has been successfully loaded into the table '{table_name}'.")
except Exception as e:
    logging.error(f"Failed to load CSV data into the table '{table_name}'.")
    logging.error(f"Error details: {e}")


A `excute_queries()` function is defined to run a list of SQL queries to compare the dimensions of the Dataframa and the PostgreSQL table.

def execute_queries(queries, db_client):
    """
    Execute a list of SQL queries and log the results.
    
    Parameters:
    - queries: List of SQL query strings.
    - db_client: DatabaseClient object containing the connection engine.
    """
    try:
        with db_client.engine.connect() as connection:
            for query in queries:
                logging.info(f"Executing query: {query}")
                # Convert query string to executable SQLAlchemy text object
                executable_query = text(query)
                result = connection.execute(executable_query)
                rows = result.fetchall()
                
                # Print and log the query results
                logging.info("Query results:")
                for row in rows:
                    logging.info(row)
    except SQLAlchemyError as e:
        logging.error("Failed to execute queries.")
        logging.error(f"Error details: {e}")

Two queries are defined in a list to count the rows and columns of the "grammys_raw" table. Then they are executed through the `excute_queries()` function.

queries = [
    "SELECT COUNT(*) FROM grammys_raw",
    "SELECT COUNT(*) FROM information_schema.columns WHERE table_name = 'grammys_raw'"
]


execute_queries(queries, db_client)

The output indicates the table has 4810 rows and 10 columns. While the method '.shape' used for the grammy's dataset Dataframe returns the same dimension, so the insertion is consistent.

df.shape

Finally, the connection to the DatabaseClient is closed through its method `.close()`.

try:
    db_client.close()
except Exception as e:
    logging.error(f"Failed to close connection to the database.")
    logging.error(f"Error details: {e}")
